Remove reference cycle in VecAccessMixin (#4033)
* Remove reference cycle in VecAccessMixin

With an associated PETSc Vec, VecAccessMixin deferred its version
property to a lambda to avoid allocating the storage until necessary.
Unfortunately, this lambda creates a reference cycle to self for all
users of the VecAccessMixin. Given that counter accesses should be
relatively infrequent, it seems fine to look up the counter type within
the method itself.

* Don't cache topological on CoordinatelessFunction

Doesn't make sense to cache a reference to self, just return self.

* Convert increment_dat_version to abstract on DataCarrier

The inheritance chain for Dat and Global puts VecAccessMixin
(rightly) behind DataCarrier. This means that by MRO, the
increment_dat_version method provided on DataCarrier will be used,
which is a null operation. I think this makes sense, given that not all
classes use this. However, because we're providing
increment_dat_version as an override through VecAccessMixin, we need to
explicitly refer to it in the inheriting classes.

* Update pyop2/types/dat.py

Otherwise updating a dat through a view doesn't increment the version.

Co-authored-by: Connor Ward <c.ward20@imperial.ac.uk>

---------

Co-authored-by: David A. Ham <david.ham@imperial.ac.uk>
Co-authored-by: Connor Ward <c.ward20@imperial.ac.uk>