A lot of progress
Current issue is that halo exchange is not occurring after computation.
I am working on cleaning up the way we track the buffers that need
exchanges and what their intents. Specifically the current challenge is
to make sure that we are thinking in terms of *buffers* and not dats
etc.