Fix docs build
We were getting undefined references due to an incompletely qualified
path and an upstream change to UFL
(https://github.com/FEniCS/ufl/pull/387/files) that is incompatible with
Sphinx (https://github.com/sphinx-doc/sphinx/issues/11225). The Sphinx
issue is still unresolved so the best course of action is to ignore the
issue for now.