Always apply entity permutations when packing
Previously this was not being respected apart from for hexes and this
meant that the cell node list was not quite the same.

This change also finally gets `_dm_renumbering` correct - and gives it a
sensible name!

Next steps are to tackle extrusion both for the periodic case and also
for facet integrals. Things are moving in the right direction.