Improve layouts for maps
If we are figuring out the subst layouts for a tree indexed with a map
then it doesn't make sense to say what the layout function is at the
root of the tree. To fix this I added a check so if there are unseen
loop indices then a NAN is used instead.